requirements.txt
python-telegram-bot==13.15
python-dotenv>=1.0.0
requests>=2.31.0
schedule>=1.2.0
Flask>=2.3.3
twelvedata==1.1.3